Add build time in upgrade preview & add default attributes

+++ plugin/src/main/java/com/eul4/rule/GenericAttribute.java
@@ -13,6 +13,6 @@ public abstract class GenericAttribute
 {
 	private Price price;
 	private int requiresTownHallLevel;
-	private long totalBuildTicks;
+	private int totalBuildTicks;
 	private Vector hologramVector;
 }
